Refactor LockType to the util module
