+++ src/frontend/connection.rs
@@ -2,8 +2,9 @@ use super::frontend_error::FrontendError;
 use super::selection::Selection;
 use super::transaction::Transaction;
 use crate::backend::{Operation, Query, Request, Response as BackendResponse};
-use crate::language::{build_statement, parse, LockType, Response, Statement};
+use crate::language::{build_statement, parse, Response, Statement};
 use crate::schema::Schema;
+use crate::util::LockType;
 use std::collections::HashMap;
 use std::io::{BufReader, BufWriter, Read, Write};
 use std::net::TcpStream;
