+++ src/language/mod.rs
@@ -23,4 +23,4 @@ pub use build_statement::build_statement;
 pub use parse_error::ParseError;
 pub use parser::parse;
 pub use response::Response;
-pub use statement::{LockType, Statement};
+pub use statement::Statement;
