+++ src/util/lock_type.rs
@@ -1,0 +1,6 @@
+#[derive(Clone)]
+pub enum LockType {
+    Read,
+    Write,
+    BlockingWrite,
+}
